requirements.txt
semver
pyyaml
requests
beautifulsoup4